% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wpx-user.R
\name{wpx_user_month_counts}
\alias{wpx_user_month_counts}
\title{Get data from the XTools User API}
\usage{
wpx_user_month_counts(username, project = "enwiki")
}
\arguments{
\item{username}{MediaWiki username.}

\item{project}{Project. Default: \code{"enwiki"} for the English Wikipedia}
}
\value{
A \code{tibble}.
}
\description{
https://www.mediawiki.org/wiki/XTools/API/User
}
\examples{
wpx_user_month_counts("Jimbo Wales")
}
